# services/facade/openapi.yaml
openapi: 3.0.0
info:
  title: VSB EInk Facade Service API
  version: 1.0.0
  description: An API for managing panels, users, and schedules in the VSB EInk Facade Service.
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HttpError:
      type: object
      properties:
        statusCode:
          type: number
        code:
          type: string
        error:
          type: string
        message:
          type: string
    UnknownType: {}
    Role:
      type: string
      enum: [ADMIN, USER]
...
    User:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkedUserGroup'
      required: [id, username, role, groups]
    InsertableUser:
      type: object
      properties:
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        password:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkableUserGroup'
      required: [username, password]
    UpdatableUser:
      type: object
      properties:
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        password:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkableUserGroup'
...
    LinkableUserGroup:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
      required: [id]
    LinkedUserGroup:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          readOnly: true
      required: [id, name]
...
  parameters:
    path:
      name: path
      in: path
      required: true
      schema:
        type: string
        default: ''
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: number
...
  responses:
    UnknownResponse:
      description: Unknown response
      content:
        '*/*':
          schema:
            $ref: '#/components/schemas/UnknownType'
  requestBodies:
    UnknownRequestBody:
      content:
        '*/*':
          schema:
            $ref: '#/components/schemas/UnknownType'
paths:
...
  /users/:
    get:
      operationId: getUsers
      tags: [users]
      security:
        - tokenAuth: [USERS_READ]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      operationId: createUser
      tags: [users]
      security:
        - tokenAuth: [USERS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertableUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      security:
        - tokenAuth: [USERS_READ]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Requested user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    patch:
      operationId: updateUser
      tags: [users]
      security:
        - tokenAuth: [USERS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatableUser'
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    delete:
      operationId: deleteUser
      tags: [users]
      security:
        - tokenAuth: [USERS_WRITE]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: User deleted
        '404':
          description: User with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /hosted/user/{path}:
    get:
      operationId: sendGetToContent
      tags: [hosted, content]
      security: []
      parameters:
        - $ref: '#/components/parameters/path'
      responses:
        default:
          $ref: '#/components/responses/UnknownResponse'
    post:
      operationId: sendPostToContent
      tags: [hosted, content]
      security: []
      parameters:
        - $ref: '#/components/parameters/path'
      requestBody:
        $ref: '#/components/requestBodies/UnknownRequestBody'
      responses:
        default:
          $ref: '#/components/responses/UnknownResponse'
    put:
      operationId: sendPutToContent
      tags: [hosted, content]
      security: []
      parameters:
        - $ref: '#/components/parameters/path'
      requestBody:
        $ref: '#/components/requestBodies/UnknownRequestBody'
      responses:
        default:
          $ref: '#/components/responses/UnknownResponse'
    patch:
      operationId: sendPatchToContent
      tags: [hosted, content]
      security: []
      parameters:
        - $ref: '#/components/parameters/path'
      requestBody:
        $ref: '#/components/requestBodies/UnknownRequestBody'
      responses:
        default:
          $ref: '#/components/responses/UnknownResponse'
    delete:
      operationId: sendDeleteToContent
      tags: [hosted, content]
      security: []
      parameters:
        - $ref: '#/components/parameters/path'
      responses:
        default:
          $ref: '#/components/responses/UnknownResponse'
...